+++ .next/server/webpack-runtime.js
@@ -125,11 +125,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-<<<<<<< HEAD
-/******/ 		__webpack_require__.h = () => ("5e293b5b0872d539")
-=======
-/******/ 		__webpack_require__.h = () => ("92d9e7696f36164c")
->>>>>>> 45ea9b083d167365b13824b153f4ecdebabe5b7d
+/******/ 		__webpack_require__.h = () => ("ac6888a8a2a77299")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/hasOwnProperty shorthand */